Add files via upload
Includes working with Datatime objects in order to know the time difference between the send dates of an email from the non-profit to the open dates of the email from the receiver which would be the UNHCR member. Did this to see to create more data points in order to increase the accuracy of our machine learning models.